requirements.txt
pyinstaller
pandas 
matplotlib 
seaborn 
numpy 
tkcalendar 
scipy
pillow
